binBoundaries ArrayList<Number> ivm state uit GWT/HTML5, ihkv DWOWIDGET-65
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/Statistiek/trunk@15460 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/fi/statistiek/frequencytable/FrequencyTableView.java
@@ -747,16 +747,16 @@ public class FrequencyTableView extends JPanel implements Observer
 				
 				String text;
 				if (type.equals(AllowedTypes.INTEGER)
-					&& (this.model.getBinBoundaries().get(1) - this.model.getBinBoundaries().get(0)) == 1)
+					&& (this.model.getBinBoundaries().get(1).doubleValue() - this.model.getBinBoundaries().get(0).doubleValue()) == 1)
 				{
 					// bin width is 1
-					text = Statistiek.getStringValue(this.model.getBinBoundaries().get(bin)); // getStringValue will give an integer value without decimals
+					text = Statistiek.getStringValue(this.model.getBinBoundaries().get(bin).doubleValue()); // getStringValue will give an integer value without decimals
 				}
 				else
 				{
-					text = Statistiek.getStringValue(this.model.getBinBoundaries().get(bin))
+					text = Statistiek.getStringValue(this.model.getBinBoundaries().get(bin).doubleValue())
 						+ " -< "
-						+ Statistiek.getStringValue(this.model.getBinBoundaries().get(bin + 1));
+						+ Statistiek.getStringValue(this.model.getBinBoundaries().get(bin + 1).doubleValue());
 				}
 				JLabel label = new JLabel(text);
